k8s.io/apiserver: move server storage code into subpackage
