+++ staging/src/k8s.io/apiserver/pkg/server/storage/storage_factory.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package storage
 
 import (
 	"fmt"
@@ -75,18 +75,6 @@ type DefaultStorageFactory struct {
 
 	// newStorageCodecFn exists to be overwritten for unit testing.
 	newStorageCodecFn func(opts StorageCodecConfig) (codec runtime.Codec, err error)
-}
-
-// StorageCodecConfig are the arguments passed to newStorageCodecFn
-type StorageCodecConfig struct {
-	StorageMediaType  string
-	StorageSerializer runtime.StorageSerializer
-	StorageVersion    schema.GroupVersion
-	MemoryVersion     schema.GroupVersion
-	Config            storagebackend.Config
-
-	EncoderDecoratorFn func(runtime.Encoder) runtime.Encoder
-	DecoderDecoratorFn func([]runtime.Decoder) []runtime.Decoder
 }
 
 type groupResourceOverrides struct {
@@ -276,59 +264,6 @@ func (s *DefaultStorageFactory) Backends() []string {
 		backends.Insert(overrides.etcdLocation...)
 	}
 	return backends.List()
-}
-
-// NewStorageCodec assembles a storage codec for the provided storage media type, the provided serializer, and the requested
-// storage and memory versions.
-func NewStorageCodec(opts StorageCodecConfig) (runtime.Codec, error) {
-	mediaType, _, err := mime.ParseMediaType(opts.StorageMediaType)
-	if err != nil {
-		return nil, fmt.Errorf("%q is not a valid mime-type", opts.StorageMediaType)
-	}
-	serializer, ok := runtime.SerializerInfoForMediaType(opts.StorageSerializer.SupportedMediaTypes(), mediaType)
-	if !ok {
-		return nil, fmt.Errorf("unable to find serializer for %q", opts.StorageMediaType)
-	}
-
-	s := serializer.Serializer
-
-	// etcd2 only supports string data - we must wrap any result before returning
-	// TODO: storagebackend should return a boolean indicating whether it supports binary data
-	if !serializer.EncodesAsText && (opts.Config.Type == storagebackend.StorageTypeUnset || opts.Config.Type == storagebackend.StorageTypeETCD2) {
-		glog.V(4).Infof("Wrapping the underlying binary storage serializer with a base64 encoding for etcd2")
-		s = runtime.NewBase64Serializer(s)
-	}
-
-	// Give callers the opportunity to wrap encoders and decoders.  For decoders, each returned decoder will
-	// be passed to the recognizer so that multiple decoders are available.
-	var encoder runtime.Encoder = s
-	if opts.EncoderDecoratorFn != nil {
-		encoder = opts.EncoderDecoratorFn(encoder)
-	}
-	decoders := []runtime.Decoder{s, opts.StorageSerializer.UniversalDeserializer()}
-	if opts.DecoderDecoratorFn != nil {
-		decoders = opts.DecoderDecoratorFn(decoders)
-	}
-
-	// Ensure the storage receives the correct version.
-	encoder = opts.StorageSerializer.EncoderForVersion(
-		encoder,
-		runtime.NewMultiGroupVersioner(
-			opts.StorageVersion,
-			schema.GroupKind{Group: opts.StorageVersion.Group},
-			schema.GroupKind{Group: opts.MemoryVersion.Group},
-		),
-	)
-	decoder := opts.StorageSerializer.DecoderToVersion(
-		recognizer.NewDecoder(decoders...),
-		runtime.NewMultiGroupVersioner(
-			opts.MemoryVersion,
-			schema.GroupKind{Group: opts.MemoryVersion.Group},
-			schema.GroupKind{Group: opts.StorageVersion.Group},
-		),
-	)
-
-	return runtime.NewCodec(encoder, decoder), nil
 }
 
 func (s *DefaultStorageFactory) ResourcePrefix(groupResource schema.GroupResource) string {
