+++ staging/src/k8s.io/apiserver/pkg/server/storage/storage_factory_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package storage
 
 import (
 	"os"
